cursors: Make build more generic/build more sizes
Breeze ships with 24 and 48 size cursors.  For common WQHD laptops,
something in between, like 40, is preferable.  Given the cursor source
is SVG, it would be possible to create arbitrary size cursors, but the
build script and xcursorgen config files hardcoded the 24 and 48 sizes.

This change turns the xcursorgen config files into something that can
be processed to generate arbitrary size cursors.  The build script has
been modified to iterate over a list of sizes that can be changed as
desired without having to change any hardcoded values.  Pointer x-y
hit values are interpolated, and where they fall between pixels, they
are rounded to the nearest pixel.  This does not seem to affect
usability, especially on the high-dpi displays for which the cursors
are intended.

This change also modifies the build script to support building multiple
similar cursors, and merges in the Breeze_Snow cursor, eliminating the
duplicate build systems between the two versions of the Breeze cursors.

Lastly, we fix a broken alias for the n-resize cursor, from
size-size_ver to size_ver.